Added functionality to detect whether a persistent file exists.
If it does not, the program will create it and notify the user, it will
then write 0 to the file to indicate there are no known users.

If it does, it opens the file in RD_ONLY mode, and then reads the first
byte into the number of users. It currently prints the number to the screen
for debugging purposes.