GIANT COMMIT OH MY GOD
This was a _huge_ restructuring of the program. A new struct
called UserRegister was added, which now holds the user list,
the number of current users, and the max number of users the list
can hold.

After looking through the code a bunch, I've tested enough to see
that all the functionality has been retained.

I believe this change will be helpful in increasing the organization
of the program.