Account Creation now supported.
The account creation process now works. If you attempt to log in
with a username that is not registered, the system will prompt you to
create an account with an account name and password.

The process concludes with a complete User object with name, hash,
and id. Once file writing is supported, this thing will really begin
to be functional.