Moved userListInit in fileSetup()
userListInit() was moved from the end of the else clause to the end
of the function, where it will be always run. This is because the
list field of the UserRegister object was not being allocated on
the initial run of the program.

While the function is meant for reading in User data from file,
running it on first use will allocate memory for the user list,
while skipping the read loop.