Complete style change to K&R
I changed all the style in the source to K&R style. This includes
next line braces for functions, underscored_names instead of camelCase,
and } else { type of style.

I believe this style looks more readable and appealing to me, and
I see it used plenty in open source projects. So I implemented it in
my own for my own experience.