Multiple organizational changes
user_list_init() is now called from main, instead of at the end of
file_setup(). It was going to be called regardless, so I thought that
it would be easier to understand if was added to the main. The warning
on interrupt string was also moved to auth.h