New users are now added to userlist
Also, a bug was fixed that was allocating 0 bytes for the
user list. It wasn't giving a seg fault because the pointer
must have been a valid address in the heap. It now allocates
16 bytes for the max number of users. The list is just an
array of User pointers, so this is sufficient.

It also checks to see that the max number of users is not exceeded
when adding new users.