Started file writing implementation
Added write_to_file(), contains an untested method of writing
all userdata stored in memory to a persistent file.

Some test areas to be aware of:

The file opening is untested, but I believe O_WRONLY and O_TRUNC
are the appropriate flags

The User assignment in the first lines of the write loop may be
destructive. This is just from a quick glance, but by incrementing
list pointer in user_reg, the actual pointer could be incremented.
Now, this might not even be a problem, seeing as the function runs
on program close, which means that the data stored in memory doesn't
need to be preserved. (Then again, this may make freeing the alloc'ed
memory more difficult.

More testing to come soon.