Bug fixes to write_to_file()
I redid the implementation in write_to_file() to make a copy of the
User **list in the UserRegister object, and then incremented it each
iteration. This allowed me to dereference the double pointer and use
the arrow notation for the deferenced User *.

I also found some bugs in the old way, where I was trying to deference
a User variable, which would obviously would have been found if I
tested more before I commit.

Some Issues:
The program will recognize the first user saved to file, but does not
recognize further unique users. However, the num_users value is saved
and updates with each user 'added'. This is either a problem in checking
whether a user exists or not or in the writing or reading of the binary
file.