User saving bug fixed
I found an old remnant of when I was only writing the num_user value
as a single byte to the binary file. Once I changed it to 4, everything
is working as it should.

Note:

I left a little bit of debugging printfs just in case I'm having
trouble with reading and writing as I start testing for more edge cases